resolved all issues described in comments above
