--- src/App.css
@@ -1,16 +1,0 @@
-/*
-.Gallery {
-  text-align: center;
-  justify-content: center;
-}
-
-
-p {
-  cursor: pointer;
-}
-
-a {
-    color: black;
-    text-decoration-line: none;
-}
-*/
