+++ src/components/theme.js
@@ -1,15 +1,10 @@
-const {createMuiTheme} = require("@material-ui/core");
-
+import {createMuiTheme} from "@material-ui/core";
 
 
 export const theme = createMuiTheme({
     overrides: {
-
-        // Style sheet name ⚛️
         MuiButton: {
-            // Name of the rule
             label: {
-                // Some CSS
                 color: 'white'
             },
         },
