+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import {Routes} from "./components/Routes";
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import {store} from './components/store'
+import {store} from './store/store'
 import { ThemeProvider } from '@material-ui/core/styles';
 import {theme} from "./components/theme";
