+++ src/store/actions/chooseBreedAction.js
@@ -1,0 +1,6 @@
+export const chooseBreedAction = (clickedBreedName) => {
+    return {
+        type: 'clickedBreed/setClickedBreed',
+        payload: clickedBreedName
+    }
+}
