+++ src/store/actions/clearImagesReducerAction.js
@@ -1,0 +1,6 @@
+export const clearImagesReducerAction = () => {
+    return {
+        type: 'imgCat/clearState',
+        payload: []
+    }
+}
