+++ src/store/reducers/breedsReducer.js
@@ -1,0 +1,7 @@
+export const breedsReducer = (state = {loading: true, breeds: []}, action) => {
+    switch (action.type) {
+        case 'breeds/loadBreeds':
+            return {loading: false, breeds: action.payload}
+        default: return state;
+    }
+}
