+++ src/store/reducers/chosenBreedReducer.js
@@ -1,0 +1,7 @@
+export const chosenBreedReducer = (state='', action) => {
+    switch (action.type) {
+        case 'clickedBreed/setClickedBreed':
+            return action.payload
+        default: return state;
+    }
+}
