+++ src/store/reducers/imagesReducer.js
@@ -1,0 +1,10 @@
+const initialState = [];
+export const imagesReducer = (state = initialState, action) => {
+    switch (action.type) {
+        case 'imgCat/getImgCat':
+            return action.payload.map((item) => item.url)
+        case 'imgCat/clearState':
+            return action.payload;
+        default: return state;
+    }
+}
