+++ src/styles/navBarStyles.js
@@ -1,0 +1,9 @@
+import {makeStyles} from "@material-ui/core";
+
+export const useStyles = makeStyles((theme) => ({
+    root: {
+        display: "flex",
+        justifyContent: "center",
+    }
+
+}))
